toy/requirements.txt
openslide-python 
Pillow
tk